Update list in API.
