+++ api/src/lib/auth/index.ts
@@ -1,16 +1,1 @@
 export { AuthConfig } from './AuthConfig';
-export { getLoggerOptions } from './AuthLogging';
-export {
-  PermissionType,
-  getDefaultPermissionSchema,
-  getPermsFromGroup,
-  type Permission,
-  type PermissionSchema,
-} from './Permissions';
-export {
-  RoleType,
-  getDefaultRoleSchema,
-  getRolesFromGroup,
-  type Role,
-  type RoleSchema,
-} from './Role';
